backend/requirements.txt
fastapi
uvicorn
pyairtable
groq
python-dotenv
pydantic
requests
